feat(component-news): replaced components-core with unity-react-core
BREAKING CHANGE: Replaced components-core with unity-react-core
