+++ packages/component-news/__mocks__/fetch-mock.js
@@ -1,10 +1,20 @@
 // @ts-check
-import fetchMock from "jest-fetch-mock";
-
+import { vi } from 'vitest';
 import * as data from "./api/feeds.json";
 
-fetchMock.enableMocks();
-fetchMock.mockResponse(() => {
+const mockFetch = vi.fn();
+
+vi.stubGlobal('fetch', mockFetch);
+
+mockFetch.mockImplementation(async () => {
   const res = { ...data };
-  return Promise.resolve(JSON.stringify(res));
+  return new Response(JSON.stringify(res), {
+    status: 200,
+    headers: { 'Content-Type': 'application/json' }
+  });
+});
+
+// Optional: Reset mocks between tests
+beforeEach(() => {
+  mockFetch.mockClear();
 });
