+++ packages/component-news/src/components/CardCarouselNews/index.jsx
@@ -1,9 +1,8 @@
 // @ts-check
-import { FeedContext } from "@asu/components-core";
+import { CardCarousel, FeedContext } from "@asu/unity-react-core";
 import React, { useContext, useEffect } from "react";
 
 import trackReactComponent from "../../../../../shared/services/componentDatalayer";
-import { CardCarousel } from "../../../../unity-react-core/src/components/ComponentCarousel/components/CardCarousel/CardCarousel";
 import { BaseFeed } from "../../core/components/BaseFeed";
 import { defaultProps } from "../../core/constants/default-props";
 import { NewsWrapper } from "./index.styles";
