+++ packages/component-news/src/components/CardGridNews/index.jsx
@@ -1,23 +1,25 @@
 // @ts-check
-import { Card, FeedContext } from "@asu/components-core";
+import { Card, feedCardButtonShape, FeedContext } from "@asu/unity-react-core";
 import React, { useContext, useEffect } from "react";
 
 import trackReactComponent from "../../../../../shared/services/componentDatalayer";
-import { feedCardButtonShape } from "../../../../components-core/src/components/FeedAnatomy/feed-prop-types";
 import { BaseFeed } from "../../core/components/BaseFeed";
 import { defaultProps } from "../../core/constants/default-props";
 import { parseInterests } from "../../core/utils";
 import { NewsWrapper } from "./index.styles";
 
 /**
- * @param {object} feed
+ *
+ * @param {Object} feed
  * @param {import("../../core/types/news-types").CardButton} cardButton
  */
-const listRow = (feed, cardButton) => (
-  <div className="card card-hover cards-items-container" key={feed.id}>
+const gridRow = (feed, cardButton) => (
+  <div
+    className="col col-12 col-md-6 col-lg-4 cards-items-container"
+    key={feed.id}
+  >
     <Card
-      type="story"
-      horizontal
+      type="default"
       eventFormat="inline"
       eventLocation={feed.location}
       clickable={!!feed.buttonLink}
@@ -28,7 +30,7 @@ const listRow = (feed, cardButton) => (
       image={feed.imageUrl}
       imageAltText={feed.title}
       linkLabel={feed.eventButtonText}
-      linkUrl={feed.eventButtonUrl || feed?.buttonLink}
+      linkUrl={feed.eventButtonUrl || feed.buttonLink}
       buttons={[
         {
           ariaLabel: cardButton.text,
@@ -47,14 +49,14 @@ const listRow = (feed, cardButton) => (
  * @param {import("../../core/types/news-types").TemplateProps} props
  */
 // eslint-disable-next-line react/prop-types
-const ListTemplate = ({ cardButton }) => {
+const GridTemplate = ({ cardButton }) => {
   const { feeds } = useContext(FeedContext); // Reading the "feeds" object from the context
 
   return (
-    <NewsWrapper className="row-spaced" data-testid="list-view-container">
+    <NewsWrapper className="row row-spaced" data-testid="grid-view-container">
       {feeds?.map((feed, index) => (
         // eslint-disable-next-line react/no-array-index-key
-        <React.Fragment key={index}>{listRow(feed, cardButton)}</React.Fragment>
+        <React.Fragment key={index}>{gridRow(feed, cardButton)}</React.Fragment>
       ))}
     </NewsWrapper>
   );
@@ -68,12 +70,12 @@ const ListTemplate = ({ cardButton }) => {
 /**
  * @param {FeedType} props
  */
-const CardListlNews = ({ cardButton, ...props }) => {
+const CardGridNews = ({ cardButton, ...props }) => {
   useEffect(() => {
     if (typeof window !== "undefined") {
       trackReactComponent({
         packageName: "component-news",
-        component: "CardListlNews",
+        component: "CardGridNews",
         type: "NA",
         configuration: {
           cardButton,
@@ -86,13 +88,16 @@ const CardListlNews = ({ cardButton, ...props }) => {
   return (
     // Calling the high order component that fetch the data
     <BaseFeed {...props}>
-      <ListTemplate
+      <GridTemplate
         cardButton={{ ...defaultProps.cardButton, ...cardButton }}
       />
     </BaseFeed>
   );
 };
 
-CardListlNews.propTypes = { ...BaseFeed.propTypes, feedCardButtonShape };
+CardGridNews.propTypes = {
+  ...BaseFeed.propTypes,
+  cardButton: feedCardButtonShape,
+};
 
-export { CardListlNews };
+export { CardGridNews };
