+++ packages/component-news/src/components/CardListlNews/index.stories.jsx
@@ -5,7 +5,7 @@ import { CardListlNews } from "./index";
 
 import { createMockParam } from "../../core/utils";
 
-/** @typedef {import("@asu/components-core/src/core/types/feed-types").FeedType } FeedType */
+/** @typedef {import("@asu/unity-react-core/src/core/types/feed-types").FeedType } FeedType */
 
 export default {
   title: "News component/Card List News",
