+++ packages/component-news/src/core/components/BaseFeed/index.jsx
@@ -3,21 +3,19 @@ import {
   FeedBody,
   FeedContainerProvider,
   FeedHeader,
-} from "@asu/components-core";
-import PropTypes from "prop-types";
-import React from "react";
-
-import {
   feedHeaderShape,
   feedCtaButtonShape,
   feedDataSourceShape,
-} from "../../../../../components-core/src/components/FeedAnatomy/feed-prop-types";
+} from "@asu/unity-react-core";
+import PropTypes from "prop-types";
+import React from "react";
+
 import { defaultProps } from "../../constants/default-props";
 import { filterData } from "../../services/data-manager";
 import { transformData } from "../../transformers/transform-data";
 
 /**
- * @typedef {import("@asu/components-core/src/core/types/feed-types").FeedType} FeedType
+ * @typedef {import("../../../../../unity-react-core/src/core/types/feed-types").FeedType} FeedType
  */
 
 /**
@@ -30,11 +28,11 @@ const BaseFeed = ({
   dataSource: pDataSource,
   maxItems,
 }) => {
-  const filters = pDataSource.filters?.replace(/_/g, " ");
+  const filters = pDataSource?.filters?.replace(/_/g, " ");
   const dataSource = { ...pDataSource, filters };
 
   // We provide in the renderBody the view specified before in the parent component, recieved as "children" in this component.
-  // We provide in the renderHeader the components-core header, if it is desired to be shown
+  // We provide in the renderHeader the unity-react-core header, if it is desired to be shown
   // We provide the maxItems prop to limit the items rendered
   // We provide the dataSource to read the url to fetch the data
   // We provide the defaultProps to use some needed default values in case they are not provided
