+++ packages/component-news/src/core/types/news-types.js
@@ -1,11 +1,11 @@
 // ts-check
 
 /**
- * @typedef {import("../../../../components-core/src/core/types/feed-types").FeedType} FeedType
+ * @typedef {import("../../../../unity-react-core/src/core/types/feed-types").FeedType} FeedType
  * */
 
 /**
- * @typedef {import("../../../../components-core/src/core/types/feed-types").FeedCardButton} CardButton
+ * @typedef {import("../../../../unity-react-core/src/core/types/feed-types").FeedCardButton} CardButton
  * */
 
 /**
