+++ packages/component-news/vitest.setup.ts
@@ -1,0 +1,9 @@
+import '@testing-library/jest-dom/vitest'
+import { cleanup } from '@testing-library/react'
+import { afterEach } from 'vitest'
+
+import "./__mocks__/fetch-mock";
+import "./__mocks__/component-carousel-mock";
+afterEach(() => {
+  cleanup()
+});
